pyproject.toml
[tool.poetry]
name = "smarttesting-python"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.17.4"
requests = "^2.27.1"
factory-boy = "^3.2.1"
injector = "^0.19.0"
pydantic = "^1.9.0"
fastapi = "^0.73.0"
wiremock = "^2.3.1"
SQLAlchemy = "^1.4.31"
docker = "^5.0.3"
psycopg2-binary = "^2.9.3"
celery = "^5.2.3"
marshmallow-dataclass = "^8.5.3"
marshmallow-enum = "^1.5.1"
kombu = "^5.2.3"
Flask = "^2.0.2"
Flask-Injector = "^0.13.0"
flask-expects-json = "^1.7.0"
pytest-docker-compose = "^3.2.1"
pymongo = "^4.0.1"
typing-inspect = "^0.7.1"
selenium = "^4.1.0"
selene = "^1.0.2"
pylint-forbidden-imports = "^1.0.0"
pact-python = "^1.5.0"
pytest-benchmark = "^3.4.1"
locust = "^2.7.3"
pybuilder = "^0.13.4"
invoke = "^1.6.0"
pytest-rerunfailures = "^10.2"
polling = "^0.3.2"
pytest-cov = "^3.0.0"
mutmut = "^2.2.0"
expects = "^0.9.0"
responses = "^0.18.0"
vcrpy = "^4.1.1"
pytest-randomly = "^3.11.0"
assertpy = "^1.1"
Flask-Admin = "^1.6.0"
Flask-Zipkin = "^0.0.5"
prometheus-client = "^0.13.1"
prometheus-flask-exporter = "^0.18.7"
freezegun = "^1.1.0"
types-freezegun = "^1.1.6"
cached-property = "^1.5.2"
importlib-metadata = "^4.11.1"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
black = "^22.1.0"
flake8 = "^4.0.1"
pylint = "^2.12.2"
isort = "^5.10.1"
pytest = "<7"
tavern = "^1.19.0"
types-requests = "^2.27.8"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
addopts = "-p no:pytest-randomly"
python_files = "test_*.py *_test.py *_tests.py"
markers = "uses_docker wiremock"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
